Prevent workspaces being created or deleted if they dont have references

+++ resources/js/collaboration.js
@@ -16,10 +16,12 @@ Statamic.$echo.booted(Echo => {
 });
 
 Statamic.$events.$on('publish-container-created', container => {
+    if (!container.reference) return;
     manager.addWorkspace(container);
     window.addEventListener('unload', () => manager.destroyWorkspace(container));
 });
 
 Statamic.$events.$on('publish-container-destroyed', container => {
+    if (!manager.workspaces[container.name]) return;
     manager.destroyWorkspace(container);
 });
